v2.0

+++ Hass.py
@@ -94,60 +94,81 @@ class Hass (object):
 #   The SimpleRPCServer class
 #   Declare method here, and client can call it directly. 
 
-    def createCluster(self, name, nodeList=[]):
+    def createCluster(self, name, nodeList=[], test=False):
         createCluster_result = recovery.createCluster(name)
         if createCluster_result["code"] == "0":
             if nodeList != []:
                 addNode_result = recovery.addNode(createCluster_result["clusterId"], nodeList)
             else :
                 addNode_result = {"code":1, "clusterId":createCluster_result["clusterId"], "message":"not add any node."}
-            try:
-                db_uuid = createCluster_result["clusterId"].replace("-","")
-                data = {"cluster_uuid":db_uuid, "cluster_name":name}
-                db.writeDB("ha_cluster", data)
-            except:
-                logging.error("Hass Hass - Access database failed.")
-                return "1;Access database failed, please wait a minute and try again."
-            if addNode_result["code"] == "0":
+            
+            if test==False:
+            #Unit test should not access database
                 try:
-                    for node in nodeList:
-                        data = {"node_name": node,"below_cluster":db_uuid}
-                        db.writeDB("ha_node", data)
-                    return "0;Create HA cluster and add computing node success, cluster uuid is %s" % createCluster_result["clusterId"]
+                    db_uuid = createCluster_result["clusterId"].replace("-","")
+                    data = {"cluster_uuid":db_uuid, "cluster_name":name}
+                    db.writeDB("ha_cluster", data)
                 except:
                     logging.error("Hass Hass - Access database failed.")
                     return "1;Access database failed, please wait a minute and try again."
+                    
+            if addNode_result["code"] == "0":
+            
+                if test==False:
+                #Unit test should not access database
+                    try:
+                        for node in nodeList:
+                            data = {"node_name": node,"below_cluster":db_uuid}
+                            db.writeDB("ha_node", data)
+                        return "0;Create HA cluster and add computing node success, cluster uuid is %s" % createCluster_result["clusterId"]
+                    except:
+                        logging.error("Hass Hass - Access database failed.")
+                        return "1;Access database failed, please wait a minute and try again."
+                        
             else:
                 return "0;The cluster is created.(uuid = "+createCluster_result["clusterId"]+") But,"+ addNode_result["message"]
+                
         else:
             return createCluster_result["code"]+";"+createCluster_result["clusterId"]
 
-    def deleteCluster(self, uuid):
+    def deleteCluster(self, uuid, test=False):
         result = recovery.deleteCluster(uuid)
-        db_uuid = uuid.replace("-","")
-        db.deleteData("DELETE FROM ha_cluster WHERE cluster_uuid = %s", db_uuid)
+        
+        if test==False:
+        #Unit test should not access database
+            db_uuid = uuid.replace("-","")
+            db.deleteData("DELETE FROM ha_cluster WHERE cluster_uuid = %s", db_uuid)
+            
         return result["code"]+";"+result["message"]
     
     def listCluster(self):
         result = recovery.listCluster()
         return result
     
-    def addNode(self, clusterId, nodeList):
+    def addNode(self, clusterId, nodeList, test=False):
         result = recovery.addNode(clusterId, nodeList)
         if result["code"] == "0":
-            try:
-                for node in nodeList:
-                    db_uuid = clusterId.replace("-", "")
-                    data = {"node_name": node,"below_cluster":db_uuid}
-                    db.writeDB("ha_node", data)
-            except:
-                return "2;System failed, please wait a minute and try again.(DB Exception)"
+        
+            if test==False:
+            #Unit test should not access database
+                try:
+                    for node in nodeList:
+                        db_uuid = clusterId.replace("-", "")
+                        data = {"node_name": node,"below_cluster":db_uuid}
+                        db.writeDB("ha_node", data)
+                except:
+                    return "2;System failed, please wait a minute and try again.(DB Exception)"
+                    
         return result["code"]+";"+result["message"]
     
-    def deleteNode(self, clusterId, nodename):
+    def deleteNode(self, clusterId, nodename, test=False):
         result = recovery.deleteNode(clusterId, nodename)
-        db_uuid = clusterId.replace("-", "")
-        db.deleteData("DELETE FROM ha_node WHERE node_name = %s && below_cluster = %s", (nodename, db_uuid))
+                
+        if test==False:
+        #Unit test should not access database
+            db_uuid = clusterId.replace("-", "")
+            db.deleteData("DELETE FROM ha_node WHERE node_name = %s && below_cluster = %s", (nodename, db_uuid))
+            
         return result["code"]+";"+result["message"]
         
     def listNode(self, clusterId) :
